lvmpolld: fix crash, fallthrough and error handling
Check for cmd_pid > 0 before sending signal.
Fix fallthrough bug and add missing NULL checks in client.
Use polling_op accessor instead of direct array index.
Fix crash in _fini on partial _init failure.
Use correct error reason for invalid poll type.

+++ daemons/lvmpolld/lvmpolld-data-utils.c
@@ -391,7 +391,8 @@ void lvmpolld_thread_data_destroy(void *thread_private)
 		 * FIXME: skip this step if lvmpolld is activated
 		 * 	  by systemd.
 		 */
-		if (!pdlv_get_polling_finished(data->pdlv))
+		if (!pdlv_get_polling_finished(data->pdlv) &&
+		    data->pdlv->cmd_pid > 0)
 			kill(data->pdlv->cmd_pid, SIGTERM);
 		pdlv_set_polling_finished(data->pdlv, 1);
 		pdst_locked_dec(data->pdlv->pdst);
